perf(api-reference, use-codemirror): lazy-load @codemirror/* behind a dynamic import boundary
